requirements.txt
# Crypto Trader Dependencies

# Exchange APIs
ccxt>=4.0.0
alpaca-trade-api>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pytz>=2023.3
